Add methods randomName and randomNumber to StringGenerator. Implements humans methods

+++ idap/CarWash/Humans/AZWasher.h
@@ -11,10 +11,10 @@
 
 @interface AZWasher : AZHuman
 
-@property (nonatomic, assign) float moneyFromCars;
+@property (nonatomic, assign) double moneyFromCars;
 
 - (void)washCar:(AZCar *)car;
-- (float)passMoneyToAccountant;
+- (double)passMoneyToAccountant;
 - (instancetype)init;
 
 @end
